feat: implement modules task
